requirements.txt
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
boto3==1.15.1
botocore==1.18.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
distlib==0.3.1
Django==3.1.1
django-crontab==0.7.1
django-ses==1.0.3
django-storages==1.10.1
facebook-sdk==2.0.0
filelock==3.0.12
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.2
importlib-metadata==1.7.0
iniconfig==1.0.1
instabot==0.117.0
mock==4.0.2
more-itertools==8.5.0
moviepy==1.0.3
mysqlclient==2.0.1
numpy==1.19.2
packaging==20.4
Pillow==7.2.0
pluggy==0.13.1
proglog==0.1.9
psycopg2-binary==2.8.5
py==1.9.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.0.2
python-magic-bin==0.4.14
pytz==2020.1
requests==2.24.0
requests-toolbelt==0.9.1
responses==0.12.0
schedule==0.6.0
scipy==1.5.2
six==1.15.0
sixer==1.6.1
sqlparse==0.3.1
toml==0.10.1
tqdm==4.49.0
tweepy==3.9.0
urllib3==1.25.10
virtualenv==20.0.31
zipp==3.1.0
